+++ src/extension.ts
@@ -18,6 +18,12 @@ export function activate(context: vscode.ExtensionContext) {
     // Init package info
     PackageInfo.init(context);
 
+    /*
+    const sc = new String('abcd');	// NOSONAR
+    // Add hover property
+    (sc as any).hoverValue = 5;
+*/
+
     // Log the extension dir
     //console.log(context.extension.id + ' folder: ' + context.extensionPath);